Conf pytest fixture, scope de fixture e arquivo conftest
