+++ libpythonpro/tests/conftest.py
@@ -1,0 +1,22 @@
+from time import sleep
+import pytest
+
+from libpythonpro.spam.db import Conexao
+
+
+@pytest.fixture(scope='module')  # Uma única instância dentro do modulo(arquivo)
+def conexao():
+    sleep(1)
+    # Setup
+    conexao_obj = Conexao()
+    yield conexao_obj
+    # Tear Down
+    conexao_obj.fechar()
+
+
+@pytest.fixture
+def sessao(conexao):
+    sessao_obj = conexao.gerar_sessao()
+    yield sessao_obj
+    sessao_obj.roll_back()
+    sessao_obj.fechar()
