+++ libpythonpro/tests/test_usuarios.py
@@ -1,24 +1,8 @@
-import pytest
-
-from libpythonpro.spam.db import Conexao
 from libpythonpro.spam.modelos import Usuario
 
 
-@pytest.fixture
-def conexao():
-    # Setup
-    conexao_obj = Conexao()
-    yield conexao_obj
-    # Tear Down
-    conexao_obj.fechar()
-
-
-@pytest.fixture
-def sessao(conexao):
-    sessao_obj = conexao.gerar_sessao()
-    yield sessao_obj
-    sessao_obj.roll_back()
-    sessao_obj.fechar()
+# @pytest.fixture(scope='function')  # Uma única instância a cada função do modulo(arquivo)
+# @pytest.fixture(scope='session')  # Uma única instância no modulo e para todos os relacionados a este.
 
 
 def test_salvar_usuario(sessao):
